ann: sample_uniform
